Added new SongExecuter Modified controller view model

+++ Soncoord.Player/ViewModels/ControllerViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using Prism.Commands;
 using Prism.Mvvm;
+using Soncoord.Business.Player;
 using Soncoord.Infrastructure.Interfaces;
 using Soncoord.Infrastructure.Interfaces.Services;
 using System;
@@ -9,24 +10,31 @@ namespace Soncoord.Player.ViewModels
     public class ControllerViewModel : BindableBase
     {
         private readonly IPlaylistService _playlistService;
-        private readonly IPlayerExecuter _executer;
+        private readonly IOutputsService _outputsService;
 
-        public ControllerViewModel(IPlaylistService playlistService, IPlayerExecuter executer)
+        public ControllerViewModel(IPlaylistService playlistService, IOutputsService outputsService)
         {
+            _outputsService = outputsService;
             _playlistService = playlistService;
-            _executer = executer;
-            _executer.IsTrackRunReverted = true;
 
-            Play = new DelegateCommand(OnPlayCommandExecute);
             Stop = new DelegateCommand(OnStopCommandExecute);
-
-            _executer.PositionChanged += PositionUpdated;
-            _executer.Started += PlayerStarted;
-            _executer.Ended += PlayerEnded;
+            Play = new DelegateCommand(OnPlayCommandExecute, OnCommandCanExecute)
+                .ObservesProperty(() => IsPlaying);
+            Next = new DelegateCommand(OnNextCommandExecute, OnCommandCanExecute)
+                .ObservesProperty(() => IsPlaying);
         }
 
         public DelegateCommand Play { get; set; }
         public DelegateCommand Stop { get; set; }
+        public DelegateCommand Next { get; set; }
+        private SongExecuter SongExecuter { get; set; }
+
+        private bool _isPlaying;
+        public bool IsPlaying
+        {
+            get => _isPlaying;
+            set => SetProperty(ref _isPlaying, value);
+        }
 
         private ISong _selectedSong;
         public ISong SelectedSong
@@ -49,12 +57,22 @@ namespace Soncoord.Player.ViewModels
             set => SetProperty(ref _totalTime, value);
         }
 
+        private bool OnCommandCanExecute()
+        {
+            return !IsPlaying;
+        }
+
         private void OnStopCommandExecute()
         {
-            _executer.Stop();
+            SongExecuter?.Stop();
         }
 
         private void OnPlayCommandExecute()
+        {
+            PlaySong();
+        }
+
+        private void OnNextCommandExecute()
         {
             PlayNextSong();
         }
@@ -71,11 +89,26 @@ namespace Soncoord.Player.ViewModels
 
         private void PlayerEnded(object sender, EventArgs e)
         {
-            _playlistService.Remove(SelectedSong);
+            UnregisterExecuter();
             PlayNextSong();
         }
 
+        private void PlayerStopped(object sender, EventArgs e)
+        {
+            UnregisterExecuter();
+        }
+
         private void PlayNextSong()
+        {
+            if (SelectedSong != null)
+            {
+                _playlistService.Remove(SelectedSong);
+            }
+
+            PlaySong();
+        }
+
+        private void PlaySong()
         {
             SelectedSong = _playlistService.GetNextSong();
             if (SelectedSong == null)
@@ -83,7 +116,29 @@ namespace Soncoord.Player.ViewModels
                 return;
             }
 
-            _executer.Play(_playlistService.GetSongSettings(SelectedSong));
+            var songSettings = _playlistService.GetSongSettings(SelectedSong);
+            var outputSettings = _outputsService.GetSettings();
+
+            SongExecuter = new SongExecuter(songSettings, outputSettings);
+            SongExecuter.PositionChanged += PositionUpdated;
+            SongExecuter.Started += PlayerStarted;
+            SongExecuter.Stopped += PlayerStopped;
+            SongExecuter.Ended += PlayerEnded;
+
+            SongExecuter.Play();
+            IsPlaying = true;
+        }
+
+        private void UnregisterExecuter()
+        {
+            SongExecuter.PositionChanged -= PositionUpdated;
+            SongExecuter.Started -= PlayerStarted;
+            SongExecuter.Stopped -= PlayerStopped;
+            SongExecuter.Ended -= PlayerEnded;
+            SongExecuter.Dispose();
+            SongExecuter = null;
+
+            IsPlaying = false;
         }
     }
 }
